gateway: no duplicate netswitch sensor ports
It's important that e.g. http://10.0.0.1/7 is not registered twice in the
system.  Otherwise it could happen that the one instance is supposed to be
turned on and the other instance to be turned off.  Because of the sync the
switch would get turned on and off all the time and it would never stop.