Merging Machines model with Netswitch model
This simplifies dealing with netswitches at multiple locations.  Especially for
auth checking and state updating the location_id is always needed.  Before this
required an extra query to machines.  This is not needed anymore.

In order to not sacrifice modularization, the netswitch related attributes get
the prefix netswitch_ and related code will stay in a file netswitch.go.  Also
one could later on move both files into models/machine/.