Set locationId cookie also in controller
In the machine view the location cookie is used, it has the path /.
In the admin view until now the locationId cookie is used, it has
the path /admin but it doesn't seem to be 100% reliable. So
additionally a locationId for / it set as well. Hopefully this will
quickfix things.