Implemented method invutil.Invoice#CreateFastbillDraft
Basically the method existed already in the monthly_earning package
but it's decoupled from the legacy stuff. For that it now has an
overwriteExisting flag. The legacy way does overwriteExisting=false,
for the new stuff .=true is used because much more synchronization
is needed.

Also some controller code got much slimmer due to the now better
architecture.